+++ plr_exercise.egg-info/PKG-INFO
@@ -1,0 +1,14 @@
+Metadata-Version: 2.1
+Name: plr-exercise
+Version: 1.0.1
+Summary: A small example package
+Home-page: UNKNOWN
+Author: Jonas Frey
+Author-email: jonfrey@ethz.ch
+License: UNKNOWN
+Platform: UNKNOWN
+Requires-Python: >=3.7
+License-File: LICENSE
+
+UNKNOWN
+
