+++ plr_exercise.egg-info/dependency_links.txt
@@ -1,0 +1,1 @@
+https://download.pytorch.org/whl/torch-2.1.0+cu121-cp38-cp38-linux_x86_64.whl
