+++ plr_exercise.egg-info/requires.txt
@@ -1,0 +1,2 @@
+numpy
+torch>=1.21
